Repo script improvements (#6)
There was a bug with the repo script where it would do fuzzy-matching.
This is not ideal because typing `repo test` when you have the
directories `test` and `test-2` should take you to the former. This has
been fixed so only exact matches work now.

Another improvement was made because previously symlinks would work, but
would cd you to the symlink itself. This is not ideal because symlinks
are nice to have a shorthand for a repo name, but after cd'ing into the
directory, it would be better to see the actual directory name. This has
also been changed so the path is resolved before cd'ing into it.